online_inference/requirements.txt
click==7.1.2
scikit-learn==0.24.1
pyyaml==3.11
marshmallow-dataclass==8.4.1
pandas==1.1.5
pytest==6.0.1
fastapi==0.63.0
uvicorn==0.13.4
requests~=2.22.0
